Stick with the default show toolbar for django-debug-toolbar
We enabled this at some point to make it easier for local development so that any IP accessing the local development server will see the django debug toolbar. - such as using a LAN IP address from a virtual machine. This seems to have gone a bit too far.

Instead we now assume that if you want to access the debug toolbar you'll always be doing it locally over 127.0.0.1.

Fixes #48.